[Feature] Add deletion functionality for old log files
When the extended logging functionality is active, huge amounts of log files may be created.  It is now possible to add a configuration attribute "log_maxage" to the plugin configuration stating the number of days after which the log files wil automatically be deleted. Deletion takes place on smarthome startup as well as once a day