[Feature] Allow delay for actions
Actions can be delayed via the new attribute "as_delay_[action_name]" in the state item.
The value of this attribute is the time to delay the action. Specify the time in seconds
or use m to specify minutes (e.g. "5" delays the action by 5 seconds, "5m" by 5 minutes)

The timer to set the delayed action is being removed if a state becomes active which also
sets this item (delayed or not). So if your delay is greather than your cycle, your
action will never be executed (except if you have set as_repeat_actions to false)!