[Task] Introduce class AbItemChild
The new class AbItemChild is base class for all classes below AbItem. It contains general methods/variables like the logger instance or the smarthome.py instance. It reduces the number of variables that have to be transferred when instancing child classes, as all these variables have their roots in the AbItem class.