[Feature] Root caller/source/item that triggered the update can be detected
If the update of the current state is triggered via eval_trigger the caller is mostly "Eval". This is not sufficient for checking the root cause for the update.
The plugin now determines what triggered an "Eval" and so traces back the "really" triggering event.
Item, caller and source are displayed in the log and can be used in conditions via the "special" conditions "original_item", "original_caller" and "original_source".
If there is no different root source for the update, the values are the same than "trigger_item", "trigger_caller" and "trigger_source"