[Feature] Add Functions to check caller/source for eval
When using an separate item to cumulate all triggers for manual changes, it may be required to ignore some changes based on caller/source if the change.
The main AutoBlind class now contains two methods that can be used for this.

    is_changed_by(self, caller, source, changed_by) returns True if the root change comes from certain callers/sources
    is_changed_by(self, caller, source, changed_by) return True if the root change comes not from certain callers/sources

"caller" and "source" are provided by eval and can just be used.
changed_by is a list of callers and sources which should be checked. The format is ["<first_caller>:<first_source>", "<next_caller>:<next:source>", ...]
callers and sources can be replaced by "*" meaning "any caller"/"any source".

Examples:

    [[[manuell]]]
    type = bool
    eval = not sh.test.rollo.manuell() if sh.return_plugin("AutoBlind").not_changed_by(caller, source, ["Init:*", "KNX:1.1.7"]) else sh.test.rollo.manuell()
    eval_trigger = test.rollo.move | test.rollo.stop

-> the value if item "test.rollo.manuell" will be inverted on every change if test.rollo.move or test.rollo.stop unless the change came from the Initialization (caller="Init", any source) or the chance came from a certain KNX device (caller="KNX", source="1.1.7" (physical address)).

In order to access the main AutoBlind-Class from eval ("sh.return_plugin("AutoBlind")"), a change in smarthome.py is required which is documented in pull request #181 on github (https://github.com/mknx/smarthome/pull/181)