[Feature] No updates are done unless startup_delay is over
Previously all triggers etc. were activated during the initialization
of an AutoBlind item. Therefore any triggering event caused the update
of the state even if the startup_delay was not over. This behavior is
now changed.

Triggers etc. are now not set during the initialization.
If "startup_delay" is a positive value, the initialization schedules
a function to be executed after the startup_delay is over. This
function initiates a first update of the state. Afterwards all triggers
are set. Any triggering events occuring before this takes place are
ignored.
if "startup_delay" is 0, the function that initates the update will be
called immediately after initialization.
if "startup_delay" is -1, only the triggers are set, no initial update
of the state will take place.