[Feature] New attribute as_name using data type AbValue to set name of state
Via the new attribute "as_name" it is possible to set the name, used
for the laststate_name-Item, by an item/eval/variable.

A new eval-function called "insert_suspend_time" calculates the end of
the suspension time and inserts this time into a text.
As first parameter this function needs the id of an item which is
changed when the suspension starts/ends. The age if this item is being
used to determine how much of the suspend time is already over.
The second parameter is the text into which the end of the suspension
should be inserted. Use srtftime/strptime format codes where the end
time of suspension should be inserted.