[Feature] Easier configuration of manual item
Three new attributes for usage with the manual item have been
introduced:

- as_manual_invert
- as_manual_include
- as_manual_exclude

With these items, the item for manual changes can be configured in a
very simple way.

as_manual_include and as_manual_exclude allow filtering caller/sources
that trigger a change of the manual item. The values of both attributes
are lists of <caller>:<source> entries, separated by "|".
If as_manual_exclude is given, the value of the manual-Item is changed
only, if the actual caller/source is not in the list given via
as_manual_exclude
If as_manual_include is given, the value of the manual-Item is changed
only, if the actual caller/source is in the list given via
as_manual_include.
If both are given, as_manual_exclude is checked first, then
as_manual_include.

In case no filtering is required, as_manual_invert can be used and set
to any value. It will cause a change of the value of the manual-Item
on every triggering of the item.