[Feature] Items can be given as relative item
Similar to the "get_item" eval function, items can be given as
relative items based on the AutoBlind object item.
To use this, the item id has to start with a ".". A single "." means
that the given item Id is  relative to the current level of the
AutoBlind object item. Every additional "." removes one level of the
AutoBlind object item before adding the item_id.

Examples (based on AutoBlind object item "my.autoblind.objectitem":
- "not.prefixed.with.dots" results to "not.prefixed.with.dots"
- ".onedot" results to "my.autoblind.objectitem.onedot"
- "..twodots" results to "my.autoblind.twodots"
- "..threedots" results to "my.threedots"
- "..threedots.further.down" results to "my.threedots.further.down"

The functionality is very similar to the eval-Function "get_item". It
will probably replace the function "get_item" in the near future.