Do not filter intended changes for AutoBlind item anymore
Changes if AutoBlind item retrigger the state calculation. For an early recalculation of the state the AutoBlind Item may be set to a value using a delay. However, as the change is triggered by the autoblind plugin, the event was filtered and no recalculation of the state was executed. With this change, the set-action uses a slight different caller when changing the according AutoBlind object item so that the filter will not block the event.