[Feature] Allow different actions for entering, staying at and leaving a state (#11)
Actions can now be placed in sub-items if the state item. The names of the subitems
are fixed and indicate when the actions will be executed:

[on_enter] -> Actions executed when entering a state (state active, previous state different)
[on_stay] -> Actions executed when staying at a state (state active same as previous state)
[on_enter_or_stay] -> Actions executed when entering or staying at a state (state active, previous state irrelevant)
[on_leave] -> Actions executed when leaving a state (just before activating the new state)

Actions directly in the state item are still possible. These actions will be added to the
"on_enter_or_stay" group and therefore behave as before.